refactor(phase-19)(pr-4a): voting/hidden_mission/accusation 디렉터리 승격 + tally.go 추출
decision 패키지의 3개 500+ 모듈 일괄 분할 (공통 테스트 헬퍼 의존성으로 원자적 이전):
- decision/voting.go (639줄) → decision/voting/ (module.go/config.go/handlers.go/reactor.go/state.go/events.go)
- decision/hidden_mission.go (559줄) → decision/hidden_mission/ (module.go/config.go/handlers.go/events.go/state.go/rules.go)
- decision/accusation.go (515줄) → decision/accusation/ (module.go/config.go/handlers.go/state.go/events.go + tally.go)

Scope C — handleAccusationVote (101줄) → tallyVotes+shouldResolve 순수 함수 추출:
- accusation/tally.go: pure function, lock 미보유. caller(handleAccusationVote)에서 m.mu.Lock() hold 유지.
- handleAccusationVote 80줄 이내로 축소.
- go test -race 통과.

decision/register.go 신설: 하위 3개 package blank-import으로 init() 보장.

공통 테스트 헬퍼(newTestDeps/testLogger/newTestEventBus)는 각 sub-package test 파일에 중복 정의 (cross-package 접근 불가로 인한 의도적 중복).

Factory 서명, engine.Register 방식, PlayerAware/PublicStateMarker 모두 불변.
F-go-3·4 해소 (500+ 파일 8→5, 함수 80+ 6→5).
