+++ apps/server/internal/module/decision/voting/reactor.go
@@ -1,0 +1,165 @@
+package voting
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"math/rand/v2"
+
+	"github.com/google/uuid"
+	"github.com/mmp-platform/server/internal/engine"
+)
+
+// VoteResult holds the outcome of a voting round.
+type VoteResult struct {
+	Results map[string]int `json:"results"`
+	Winner  string         `json:"winner"`
+	IsTie   bool           `json:"isTie"`
+	Round   int            `json:"round"`
+}
+
+// --- PhaseReactor ---
+
+// ReactTo handles engine-driven lifecycle actions (open / close voting).
+func (m *VotingModule) ReactTo(_ context.Context, action engine.PhaseActionPayload) error {
+	switch action.Action {
+	case engine.ActionOpenVoting:
+		return m.openVoting(action.Params)
+	case engine.ActionCloseVoting:
+		return m.closeVoting()
+	default:
+		return fmt.Errorf("voting: unsupported action %q", action.Action)
+	}
+}
+
+// SupportedActions reports which phase actions this module reacts to.
+func (m *VotingModule) SupportedActions() []engine.PhaseAction {
+	return []engine.PhaseAction{
+		engine.ActionOpenVoting,
+		engine.ActionCloseVoting,
+	}
+}
+
+type openVotingParams struct {
+	TotalPlayers int `json:"totalPlayers"`
+	AlivePlayers int `json:"alivePlayers"`
+}
+
+func (m *VotingModule) openVoting(params json.RawMessage) error {
+	m.mu.Lock()
+
+	if params != nil && len(params) > 0 {
+		var p openVotingParams
+		if err := json.Unmarshal(params, &p); err != nil {
+			m.mu.Unlock()
+			return fmt.Errorf("voting: invalid OPEN_VOTING params: %w", err)
+		}
+		m.totalPlayers = p.TotalPlayers
+		m.alivePlayers = p.AlivePlayers
+	}
+
+	m.isOpen = true
+	m.votes = make(map[uuid.UUID]string)
+	m.currentRound++
+
+	round := m.currentRound
+	maxRounds := m.config.MaxRounds
+	m.mu.Unlock()
+
+	m.deps.EventBus.Publish(engine.Event{
+		Type: "vote.opened",
+		Payload: map[string]any{
+			"round":    round,
+			"maxRound": maxRounds,
+		},
+	})
+	return nil
+}
+
+func (m *VotingModule) closeVoting() error {
+	m.mu.Lock()
+
+	if !m.isOpen {
+		m.mu.Unlock()
+		return fmt.Errorf("voting: voting is not open")
+	}
+	m.isOpen = false
+
+	result := m.tallyResults()
+	m.lastResult = &result
+	m.mu.Unlock()
+
+	m.deps.EventBus.Publish(engine.Event{
+		Type:    "vote.result",
+		Payload: result,
+	})
+	return nil
+}
+
+// tallyResults counts votes, checks participation, and handles ties.
+// Must be called under write lock.
+func (m *VotingModule) tallyResults() VoteResult {
+	counts := make(map[string]int)
+	for _, target := range m.votes {
+		if target != "" { // skip abstentions in count
+			counts[target]++
+		}
+	}
+
+	result := VoteResult{
+		Results: counts,
+		Round:   m.currentRound,
+	}
+
+	// Check minimum participation.
+	eligible := m.alivePlayers
+	if m.config.DeadCanVote {
+		eligible = m.totalPlayers
+	}
+	if eligible > 0 {
+		participation := (len(m.votes) * 100) / eligible
+		if participation < m.config.MinParticipation {
+			result.IsTie = true
+			result.Winner = ""
+			return result
+		}
+	}
+
+	// Find the winner(s).
+	maxVotes := 0
+	var winners []string
+	for target, count := range counts {
+		if count > maxVotes {
+			maxVotes = count
+			winners = []string{target}
+		} else if count == maxVotes {
+			winners = append(winners, target)
+		}
+	}
+
+	if len(winners) == 0 {
+		result.IsTie = true
+		result.Winner = ""
+		return result
+	}
+
+	if len(winners) == 1 {
+		result.Winner = winners[0]
+		result.IsTie = false
+		return result
+	}
+
+	// Tie handling.
+	result.IsTie = true
+	switch m.config.TieBreaker {
+	case "random":
+		result.Winner = winners[rand.IntN(len(winners))]
+		result.IsTie = false
+	case "no_result":
+		result.Winner = ""
+	case "revote":
+		// Remains a tie; the engine should call OPEN_VOTING again if under maxRounds.
+		result.Winner = ""
+	}
+	return result
+}
