+++ apps/server/internal/module/decision/voting/voting_test.go
@@ -1,4 +1,4 @@
-package decision
+package voting
 
 import (
 	"context"
